+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## New version
 - Fix session provisioning timeout and delay handling
+- Add on_schema_change possibility
+- Fix table materialization for Delta models
+- Change GlueColumn parent from base Column to SparkColumn
 - Fix get_columns_in_relation function to stop returning additional partition columns
 
 ## v1.8.1